Download model files in parallel when < X files
WIP: Missing test and some refactor to make the code more reasable

http://b/341160276

+++ src/kagglehub/http_resolver.py
@@ -14,7 +14,10 @@ from kagglehub.clients import KaggleApiV1Client
 from kagglehub.handle import ModelHandle
 from kagglehub.resolver import Resolver
 
+from tqdm.contrib.concurrent import thread_map
+
 MODEL_INSTANCE_VERSION_FIELD = "versionNumber"
+MAX_NUM_FILES_DIRECT_DOWNLOAD = 25
 
 logger = logging.getLogger(__name__)
 
@@ -45,28 +48,47 @@ class ModelHttpResolver(Resolver[ModelHandle]):
             os.makedirs(os.path.dirname(out_path), exist_ok=True)
             api_client.download_file(url_path + "/" + path, out_path, h)
         else:
-            # Downloading the full archived bundle.
-            archive_path = get_cached_archive_path(h)
-            os.makedirs(os.path.dirname(archive_path), exist_ok=True)
+            # List the files and decide how to download them:
+            # - <= 25 files: Download files in parallel
+            # > 25 files: Download the archive and uncompress
+            (files, has_more) = _list_files(api_client, h)
+            if has_more:
+                # Downloading the full archived bundle.
+                archive_path = get_cached_archive_path(h)
+                os.makedirs(os.path.dirname(archive_path), exist_ok=True)
 
-            # First, we download the archive.
-            api_client.download_file(url_path, archive_path, h)
+                # First, we download the archive.
+                api_client.download_file(url_path, archive_path, h)
 
-            # Create the directory to extract the archive to.
-            os.makedirs(out_path, exist_ok=True)
+                # Create the directory to extract the archive to.
+                os.makedirs(out_path, exist_ok=True)
 
-            if not tarfile.is_tarfile(archive_path):
-                msg = "Unsupported archive type."
-                raise ValueError(msg)
+                if not tarfile.is_tarfile(archive_path):
+                    msg = "Unsupported archive type."
+                    raise ValueError(msg)
 
-            # Extract all files to this directory.
-            logger.info("Extracting model files...")
-            with tarfile.open(archive_path) as f:
-                # Model archives are created by Kaggle via the Databundle Worker.
-                f.extractall(out_path)
+                # Extract all files to this directory.
+                logger.info("Extracting model files...")
+                with tarfile.open(archive_path) as f:
+                    # Model archives are created by Kaggle via the Databundle Worker.
+                    f.extractall(out_path)
 
-            # Delete the archive
-            os.remove(archive_path)
+                # Delete the archive
+                os.remove(archive_path)
+            else:
+                # Download files individually in parallel
+                def _inner_download_file(file: str):
+                    file_out_path = out_path + "/" + file 
+                    os.makedirs(os.path.dirname(file_out_path), exist_ok=True)
+                    api_client.download_file(url_path + "/" + file, file_out_path, h)
+
+                thread_map(
+                    _inner_download_file,
+                    files,
+                    desc=f"Downloading {len(files)} files",
+                    max_workers=8 # Never use more than 8 threads in parallel to download files.
+                )
+                    
 
         mark_as_complete(h, path)
         return out_path
@@ -80,6 +102,20 @@ def _get_current_version(api_client: KaggleApiV1Client, h: ModelHandle) -> int:
 
     return json_response[MODEL_INSTANCE_VERSION_FIELD]
 
+def _list_files(api_client: KaggleApiV1Client, h: ModelHandle) -> tuple[list[str], bool]:
+    json_response = api_client.get(_build_list_model_instance_version_files_url_path(h), h)
+    if "files" not in json_response:
+        msg = f"Invalid ListModelInstanceVersionFiles API response. Expected to include a 'files' field"
+        raise ValueError(msg) 
+    
+    files = []
+    for f in json_response['files']:
+        files.append(f['name'])
+    
+    has_more = "nextPageToken" in json_response and json_response["nextPageToken"] != ""
+    
+    return (files, has_more)
+
 
 def _build_get_instance_url_path(h: ModelHandle) -> str:
     return f"models/{h.owner}/{h.model}/{h.framework}/{h.variation}/get"
@@ -87,3 +123,6 @@ def _build_get_instance_url_path(h: ModelHandle) -> str:
 
 def _build_download_url_path(h: ModelHandle) -> str:
     return f"models/{h.owner}/{h.model}/{h.framework}/{h.variation}/{h.version}/download"
+
+def _build_list_model_instance_version_files_url_path(h: ModelHandle):
+    return f"models/{h.owner}/{h.model}/{h.framework}/{h.variation}/{h.version}/files?page_size={MAX_NUM_FILES_DIRECT_DOWNLOAD}"
